pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["GunkinRoman <dj89611804003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
oauth2 = "^1.9.0.post1"
sqlalchemy = "^2.0.20"
asyncpg = "^0.28.0"
psycopg2 = "^2.9.8"
asyncio = "^3.4.3"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
